third_party/WebKit/LayoutTests/fast/css-grid-layout/named-grid-line-get-set.html

<!DOCTYPE HTML>
<html>
<head>
<script>
if (window.testRunner)
    testRunner.overridePreference("WebKitCSSGridLayoutEnabled", 1);
</script>
<link href="resources/grid.css" rel="stylesheet">
<style>
.gridWithFixed {
    grid-definition-columns: "first" 10px;
    grid-definition-rows: "first" 15px;
}
.gridWithPercent {
    grid-definition-columns: 53% "last";
    grid-definition-rows: 27% "last";
}
.gridWithAuto {
    grid-definition-columns: "first" auto;
    grid-definition-rows: auto "last";
}
.gridWithMinMax {
    grid-definition-columns: "first" minmax(10%, 15px);
    grid-definition-rows: minmax(20px, 50%) "last";
}
.gridWithFixedMultiple {
    grid-definition-columns: "first" "nav" 10px "last";
    grid-definition-rows: "first" "nav" 15px "last";
}
.gridWithPercentageSameStringMultipleTimes {
    grid-definition-columns: "first" "nav" 10% "nav" 15% "last";
    grid-definition-rows: "first" "nav2" 25% "nav2" 75% "last";
}
</style>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<div class="grid gridWithFixed" id="gridWithFixedElement"></div>
<div class="grid gridWithPercent" id="gridWithPercentElement"></div>
<div class="grid gridWithAuto" id="gridWithAutoElement"></div>
<div class="grid gridWithMinMax" id="gridWithMinMax"></div>
<div class="grid gridWithFixedMultiple" id="gridWithFixedMultiple"></div>
<div class="grid gridWithPercentageSameStringMultipleTimes" id="gridWithPercentageSameStringMultipleTimes"></div>
<script>
    description('Test that setting and getting grid-definition-columns and grid-definition-rows works as expected');

    function testValue(gridElement, namedGridDefinitionColumns, namedGridDefinitionRows)
    {
        this.gridElement = gridElement;
        shouldBeEqualToString("getComputedStyle(gridElement, '').getPropertyValue('grid-definition-columns')", namedGridDefinitionColumns);
        shouldBeEqualToString("getComputedStyle(gridElement, '').getPropertyValue('grid-definition-rows')", namedGridDefinitionRows);
    }

    function testCSSValue(gridElementId, namedGridDefinitionColumns, namedGridDefinitionRows)
    {
        testValue(document.getElementById(gridElementId), namedGridDefinitionColumns, namedGridDefinitionRows);
    }

    debug("Test getting grid-definition-columns and grid-definition-rows set through CSS");
    testCSSValue("gridWithFixedElement", "first 10px", "first 15px");
    testCSSValue("gridWithPercentElement", "53% last", "27% last");
    testCSSValue("gridWithAutoElement", "first auto", "auto last");
    testCSSValue("gridWithMinMax", "first minmax(10%, 15px)", "minmax(20px, 50%) last");
    testCSSValue("gridWithFixedMultiple", "first nav 10px last", "first nav 15px last");
    testCSSValue("gridWithPercentageSameStringMultipleTimes", "first nav 10% nav 15% last", "first nav2 25% nav2 75% last");

    debug("");
    debug("Test getting and setting grid-definition-columns and grid-definition-rows through JS");
    var element = document.createElement("div");
    document.body.appendChild(element);
    element.style.gridDefinitionColumns = "'first' 18px";
    element.style.gridDefinitionRows = "66px 'last'";
    testValue(element, "first 18px", "66px last");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.gridDefinitionColumns = "'first' 55%";
    element.style.gridDefinitionRows = "40% 'last'";
    testValue(element, "first 55%", "40% last");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.gridDefinitionColumns = "'first' auto";
    element.style.gridDefinitionRows = "auto 'last'";
    testValue(element, "first auto", "auto last");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.gridDefinitionColumns = "'first' min-content";
    element.style.gridDefinitionRows = "min-content 'last'";
    testValue(element, "first min-content", "min-content last");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.gridDefinitionColumns = "'first' max-content";
    element.style.gridDefinitionRows = "max-content 'last'";
    testValue(element, "first max-content", "max-content last");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.gridDefinitionColumns = "'first' minmax(55%, 45px)";
    element.style.gridDefinitionRows = "minmax(30px, 40%) 'last'";
    testValue(element, "first minmax(55%, 45px)", "minmax(30px, 40%) last");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.font = "10px Ahem";
    element.style.gridDefinitionColumns = "'first' minmax(22em, max-content)";
    element.style.gridDefinitionRows = "minmax(max-content, 5em) 'last'";
    testValue(element, "first minmax(220px, max-content)", "minmax(max-content, 50px) last");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.font = "10px Ahem";
    element.style.gridDefinitionColumns = "'first' minmax(22em, max-content)";
    element.style.gridDefinitionRows = "minmax(max-content, 5em) 'last'";
    testValue(element, "first minmax(220px, max-content)", "minmax(max-content, 50px) last");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.gridDefinitionColumns = "'first' minmax(min-content, max-content)";
    element.style.gridDefinitionRows = "minmax(max-content, min-content) 'last'";
    testValue(element, "first minmax(min-content, max-content)", "minmax(max-content, min-content) last");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.gridDefinitionColumns = "'first' 'nav' minmax(min-content, max-content) 'last'";
    element.style.gridDefinitionRows = "'first' 'nav' minmax(max-content, min-content) 'last'";
    testValue(element, "first nav minmax(min-content, max-content) last", "first nav minmax(max-content, min-content) last");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.gridDefinitionColumns = "'first' 'nav' minmax(min-content, max-content) 'nav' auto 'last'";
    element.style.gridDefinitionRows = "'first' 'nav2' minmax(max-content, min-content) 'nav2' minmax(10px, 15px) 'last'";
    testValue(element, "first nav minmax(min-content, max-content) nav auto last", "first nav2 minmax(max-content, min-content) nav2 minmax(10px, 15px) last");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.gridDefinitionColumns = "'foo' 'bar' auto 'foo' auto 'bar'";
    element.style.gridDefinitionRows = "'foo' 'bar' auto 'foo' auto 'bar'";
    testValue(element, "foo bar auto foo auto bar", "foo bar auto foo auto bar");

    debug("");
    debug("Test getting and setting invalid grid-definition-columns and grid-definition-rows through JS");
    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.gridDefinitionColumns = "'foo'";
    element.style.gridDefinitionRows = "'bar";
    testValue(element, "none", "none");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.gridDefinitionColumns = "'foo' 'bar'";
    element.style.gridDefinitionRows = "'bar' 'foo'";
    testValue(element, "none", "none");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
